u done, strings done, float with # missing

+++ utils/ft_putnbr.c
@@ -12,7 +12,7 @@
 
 #include "ft_printf.h"
 
-void	ft_putnbr(int n)
+void	ft_putnbr(intmax_t n)
 {
 	long int a;
